h-7 some changes
done
